ci: [NGOv1.X] Corrections of CI setup issues (#3326)
This PR focuses on fixing CI issues raised in
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3303
and before. Specifically:

1. 2023.3 editor was removed from supported list since based on official
info it transitioned into 6000 editor and 2023 editor (by itself) is not
supported. The associated ticket was (MTT-11355)
2. In case of ubuntu and 2021 editor (MTT-11354) correct version of
libssl is being installed
3. UTR command was corrected to use preinstalled UTR on the image
instead of downloading it
4. Removed upm-ci project pack definition and converted Code Coverage
test to use upm-pvp pack job
5. Simplifies some unnecessary CI commands
6. Added detailed comments regarding CI setup and commands
7. Created CI_README file which describes overall configuration
structure.
8. For macOS devices we now use M1 model (Apple Silicon) in contrast to
default x64 Intel Mac which introduced a lot of bit-flipping errors

---------

Co-authored-by: NoelStephensUnity <noel.stephens@unity3d.com>