ci: [1.X] disabling WebGL Build job on macOS due to "Light baking could not be started because no valid OpenCL device could be found" error. (#3370)
Backport of
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3369

After switching from Intel macOS to Apple Silicon due to the common
problem with bit-flipping on Intel macOS the WebGL build job started
failing due to "Light baking could not be started because no valid
OpenCL device could be found" error.
We need to figure out if this can be fixed or how to proceed with it.
* PR that switched macOS device -->
[NGOv1.X](https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3326),
[NGOv2.X](https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3329)

The issue exists on both NGO branches (*develop* and *develop-2.0.0*) so
the solution needs to be applied on both. For now I'm disabling this job
and tracking it in MTT-11726