fix: In-Scene Placed Object Parenting and Serialization Order (back port) (#3388)
This is the back port of #3387 that resolves some issues with in-scene
placed NetworkObjects in parent-child hierarchy and the synchronization
of late joining clients when the hierarchy has changed from that of its
default hierarchy defined within the scene.


[MTT-11883](https://jira.unity3d.com/browse/MTT-11883)

## Changelog

- Fixed: Issue where in-scene placed `NetworkObjects` could fail to
synchronize its transform properly (especially without a
`NetworkTransform`) if their parenting changes from the default when the
scene is loaded and if the same scene remains loaded between network
sessions while the parenting is completely different from the original
hierarchy.
- Changed: The scene loading event serialization order for in-scene
placed `NetworkObject`s to be based on their parent-child hierarchy.



## Testing and Documentation

- Includes no additional tests (_requires manual testing due to scene
loading constraints_).
- No documentation changes or additions were necessary.