ci: Improving trigger jobs coverage (#3408)
This PR aims to fix an oversight of having invalid dependency in Nightly
trigger as well as adding more variations to PR trigger job.

Additionally it introduces _**run_quick_checks**_ job that was added as
a requirement for package_pack and project_pack jobs. It will run as a
first check (also on PR trigger) to catch any basic errors (PVP and
project standards) and ONLY IF this job will pass, the rest of PR
trigger job dependencies will be executed.
This way provides faster feedback loop for developers and allow to
minimize resource usage when unnecessary

Failing result may look like this

![image](https://github.com/user-attachments/assets/bc89acc7-1fad-4330-92df-407bef45c308)
Where we can see that many jobs failed but in reality we can see that
only PVP jobs failed and others were never executed and failed due to
their dependencies

## Backport
This is specific change to develop branch but it's somewhat reflected in
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3409