ci: Changed standalone tests trigger on PRs to use only Ubuntu (#3432)
It was noted that after we switched macOS from x64 to Apple Silicon the
distribution times started taking up to 40m which is fine for Nightly
jobs but it's frustrating for PRs triggers.
We decided that coverage on Ubuntu is enough for PR trigger since all
other standalone desktop devices are being tested during Nightly.

## Backport
Port for `develop-2.0.0` branch is present in
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3431