fix: authority changing ownership via rpc can result in the same previous and current clientid values [BackPort] (#3434)
This fixes an issue with the synchronization of NetworkVariables when
changing ownership where it was possible to have the previous owner be
equal to the current owner when a client used an RPC to change
ownership.

This includes an additional check for dirty states of any collections
base NetworkVariables.

## Changelog

- Fixed: Issue where when a client changes ownership via RPC the
`NetworkBehaviour.OnOwnershipChanged` can result in identical previous
and current owner identifiers.

## Testing and Documentation

- Includes integration test
`NetworkObjectOwnershipTests.TestAuthorityChangingOwnership`.
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport

This is a backport of #3347.