feat: Add support for connecting via a hostname instead of IP [MTT-10139] (#3440)
Add support for connecting via a hostname instead of IP when using
supported versions of unity and unity transport.

## Changelog

- Added: When using UnityTransport >=2.4 and Unity >= 6000.1.0a1,
`SetConnectionData` will accept a fully qualified hostname instead of an
IP as a connect address on the client side.

## Testing and Documentation

- No tests have been added. (Manual tests were performed. The CI
environment doesn't have "localhost" consistently defined to be able to
run automated tests.)
- Includes edits to existing public API documentation.

## Backport

Backport of
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3441

---------

Co-authored-by: NoelStephensUnity <noel.stephens@unity3d.com>