fix: NetworkVariables with NetworkVariableUpdateTraits can cause other NetworkVariables to drop changes (#3465)
This PR resolves an issue where a `NetworkBehaviour` with multiple
`NetworkVariables` could drop changes if one of the `NetworkVariables`
has unique `NetworkVariableUpdateTraits` set, is dirty, but is not ready
to send.

Based on user @khyperia's submission #3462.

## Changelog

- Fixed: issue where `NetworkVariable`s on a `NetworkBehaviour` could
fail to synchronize changes if one has `NetworkVariableUpdateTraits` set
and is dirty but is not ready to send.

## Testing and Documentation

- Includes integration test
`NetworkVariableTraitsTests.WhenNonTraitsIsDirtyButTraitsIsNotReadyToSend`.
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport

This required and up-port to v2.x (#3466)

---------

Co-authored-by: khyperia <953151+khyperia@users.noreply.github.com>