fix: dont destroy with owner not being honored when client exits [Test Backport of 3477] (#3478)
The v1.x branch does not have the same issue as #3477. This PR is only
the backport of the test added to 3477 to assure we have the same kind
of coverage in both versions.

## Changelog

NA

## Testing and Documentation

- Includes integration test
`NetworkObjectDontDestroyWithOwnerTests.NetworkShowThenClientDisconnects`.
- No documentation changes or additions were necessary.


<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport
This is a test-only backport of #3477.
<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->