docs: [1.X] fixes of PVP exceptions (round 2) (#3374)
This PR is a continuation of
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3300
It focuses on fixing PVP errors related to PVP-151-1 "**_Public APIs
should be documented_**" and all remaining PVP errors **present in
pvpExceptions.json file**

We have 87 errors left from PVP-151-1 at the time of this PR creation.
If anyone want's to add to it what we should do is:

1. Check errors present in pvpExceptions.json file (preferably under
"PVP-150-1", choose some to fix
2. Address those errors and when making commit remove the fixed errors
from pvpExceptions file
The first commit of this PR is an example of how such process should
look like

In general [new CI
implementation](https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3193)
will catch all new errors like this so all we need to do is fix the
present ones. This PR **does not** focus on PVP present in gold profile
but if capacity allows, those can also be addressed.

**Errors present in "PVP-41-1", "PVP-300-4" are to be expected** so
there is no need of fixing this one!
Another thing is that we could ignore error related to missing APIs for
tests (so TestHelpers, EditorTests, RuntimeTests) because those were
made internal in 2.X version (that's why those errors are not popping by
there). This does not affect users so we don't need to add anything
there

## Backport 
This has its equivalent in
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3375

---------

Co-authored-by: Emma <emma.mcmillan@unity3d.com>
Co-authored-by: Noel Stephens <noel.stephens@unity3d.com>