ci: [Backport] Fixing WebGL timeout and reenabling macOS job (#3511)
This PR is a backport of
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3507

This PR focuses on investigating often timeouts of WebGL job happening
on Ubuntu platform (note that there are no timeouts on Windows) and also
on re-enabling macOS job that was disabled due to error "_Light baking
could not be started because no valid OpenCL device could be found_" and
is tracked in MTT-11726

When it comes to the error on macOS, the build was failing due to a
light baking error in different scenes with message "_no valid OpenCL
device could be found._" This common issue on Apple Silicon VMs, where
OpenCL support is often missing for the GPU lightmapper, can be resolved
by disabling "Baked Global Illumination" for the scene. Since those
scenes in testproject do not really need baked lighting and I didn't
spot any difference with this option disabled I went forward with
disabling it for testproject scenes and this resolves macOS issue.
**Note** that most changes are related to testproject updates that
happened automatically when I changed the mentioned lighting setting

For the Ubuntu timeout problem I restored image version to v4 instead of
pointing to specific version since the issue that caused us to do so was
resolved. I also increased flavor to xlarge (from large) because it
seems that with more resources the process is not hanging anymore (will
run test 5 times to confirm)

## Backport

https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3507