chore: Reduce the number of jobs running on PR triggers (#3543)
Reduce the number of jobs that are running on each PR from 43 jobs to
26.

## Changelog

- Changed: Updated the automated test configuration.

## Testing and Documentation


## Backport

<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->

Up ported by #3545

---------

Co-authored-by: michalChrobot <michal.chrobot@unity3d.com>