fix: connection approval cannot exceed mtu size - backport (3564) (#3565)
This PR resolves the issue where a NetworkConfig.ConnectionData that
caused the ConnectionRequestMessage to exceed the transport's MTU size
would result in a buffer overflow error by updating
ConnectionRequestMessage to use the ReliableFragmentedSequenced pipeline
as opposed to the ReliableSequenced pipeline.

[MTTB-1467](https://jira.unity3d.com/browse/MTTB-1467)

## Changelog

- Fixed: issue where NetworkConfig.ConnectionData could cause the
ConnectionRequestMessage to exceed the transport's MTU size and would
result in a buffer overflow error.

## Testing and Documentation

- Includes modifications to `ConnectionApproval` integration tests to
use a > MTU size `NetworkConfig.ConnectionData`.
- No documentation changes or additions were necessary.

## Backport

This is a back port of #3564.